Fixed deleted message - filename ?

+++ back-end/controllers/message.js
@@ -9,14 +9,20 @@ exports.createMessage = (req, res, next) => {
     const title = req.body.title
     const content = req.body.content
     const userId = req.body.userId
-    const attachment = `${req.protocol}://${req.get('host')}/images-mess/${req.file.filename}`
 
-    const newMessage = {
+
+    const newMessage = req.file ? {
         userId: userId,
         title: title,
         content: content,
-        attachment: attachment
+        attachment: `${req.protocol}://${req.get('host')}/images-mess/${req.file.filename}`
+    } : {
+        userId: userId,
+        title: title,
+        content: content,
+        attachment: null
     }
+
     try {
         models.Message.create(newMessage)
             .then((newMessage) => res.status(201).json({ newMessage, message: 'Message created !' }))
@@ -69,8 +75,8 @@ exports.deleteMessage = (req, res, next) => {
             where: { id: req.params.id }
         })
             .then((message) => {
-                if (userId == message.userId) {
-                    const filename = message.attachment.split('/images-mess/')[1]
+                const filename = req.file ? message.attachment.split('/images-mess/')[1] : null
+                if (userId == message.userId && filename !== null) {
                     fs.unlink(`images-mess/${filename}`, () => {
                         models.Message.destroy({
                             where: { id: req.params.id }
@@ -78,11 +84,17 @@ exports.deleteMessage = (req, res, next) => {
                             .then(() => res.status(200).json({ message: 'Message Deleted !' }))
                             .catch((error) => res.status(400).json({ error }))
                     })
+                } else if (userId == message.userId && filename === null) {
+                    models.Message.destroy({
+                        where: { id: req.params.id }
+                    })
+                        .then(() => res.status(200).json({ message: 'Message Deleted !' }))
+                        .catch((error) => res.status(400).json({ error }))
                 } else {
                     res.status(401).json({ message: 'You can\'t delete this message, userId not match !' })
                 }
             })
-            .catch((error) => res.status(404).json({ error }))
+            .catch((error, text) => res.status(404).json({ error }))
     } catch (error) {
         res.status(500).json({ error })
     }
